whoop whoop i am committing a version that works even if it's a mess just so we have some place to return.
Added slugs_controller.rb because we'll need it eventually anyway. Tried to make it so all the slug and ID confusion gets resolved on one request pass through the thing, and then afterwards slugs work the same as post ids. (assuming the same slug isn't in the same thread which would probably be a problem)